backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0
numpy<2.0.0
opencv-python-headless==4.8.1.78
requests>=2.31.0
requests[socks]>=2.31.0
PySocks>=1.7.1
redis>=5.0.0